Move record builder into core_python.pkg
Move and consolidate the record builder into swanlab/sdk/internal/core_python/pkg/builder. Updated imports across core_python modules to use the new builder instead of the old RB alias, and adjusted metrics to import the core_python builder. Exposed builder in core_python.pkg.__init__ and removed the legacy swanlab/sdk/internal/pkg/builder module and the probe_python/builder. Added helper functions build_resume_column and build_auto_column in the new builder, and updated package __all__ and tests path to reflect the relocation.
