+++ swanlab/sdk/internal/core_python/pkg/__init__.py
@@ -5,6 +5,6 @@
 @description: core python 工具包
 """
 
-from . import counter, executor
+from . import builder, counter, executor
 
-__all__ = ["executor", "counter"]
+__all__ = ["executor", "counter", "builder"]
